Move preroller to it's own class
